validation a clinic cases, mostrar errores en los campos, y mensaje de confirmacion al borrar un clinic case en index

+++ public/js/script.js
@@ -20,27 +20,29 @@ $(document).ready(function() {
         var id = $(this).data("id");
         var token = $(this).data("token");
         var tr = $(this).closest("tr");
-        $.ajax(
-            {
-                type: "DELETE",
-                url: "./clinic-case/delete/" + id,
-                cache: false,
-                data: {
-                    "_token": token
-                },
-                success: function () {
-                    tr.fadeOut(1000, function(){
-                        $(this).remove();
-                    });
-                    console.log("it Work");
-                }
-            });
-
-        console.log("It failed");
+        var x = confirm("Are you sure you want to delete?");
+        if (x)
+            $.ajax(
+                {
+                    type: "DELETE",
+                    url: "./clinic-case/delete/" + id,
+                    cache: false,
+                    data: {
+                        "_token": token
+                    },
+                    success: function () {
+                        tr.fadeOut(500, function(){
+                            $(this).remove();
+                        });
+                        console.log("it Work");
+                    }
+                });
+        else
+            console.log("It failed");
     });
 
 
-	// Main Menu
+    // Main Menu
 	$('#main-nav').affix({
 		offset: {
 			top: $('header').height()
